requirements.txt
python-docx
pdfplumber
scikit_learn
streamlit
nltk
